requirements.txt
--extra-index-url https://test.pypi.org/simple/
# Core dependencies
fastapi==0.104.1
streamlit==1.31.1
transformers==4.35.2
python-dotenv==1.0.0
pymupdf==1.23.7
ebooklib==0.18
beautifulsoup4==4.12.2
accelerate==0.25.0
scipy==1.11.4  # Required for bitsandbytes, using newer version with pre-built wheels
bitsandbytes==0.39.0  # Latest version from test.pypi.org that supports CPU quantization
nltk==3.8.1  # Added for sentence tokenization
psutil==5.9.0  # For memory monitoring
pytest==8.2.0  # For testing
pytest-mock==3.12.0  # For mocking in tests
pytest-cov==5.0.0    # For code coverage
python-docx==1.1.0  # For Word document generation
reportlab==4.1.0    # For PDF generation
sentencepiece
